requirements.txt
pyjokes
wikipedia
pyttsx3
SpeechRecognition
google-generativeai

